LayoutTests/webrtc/video-rotation-expected.txt

  

PASS Setting video exchange 
PASS Track is enabled, video should not be black 
PASS Track is enabled and rotated, video should not be black and should change size 
